Finish documenting (I hope)
