--- task_2_3_1/src/main/java/kozlov/kirill/snake/view_model/SceneManagerAccessible.java
@@ -1,9 +1,0 @@
-package kozlov.kirill.snake.view_model;
-
-import kozlov.kirill.snake.ExcludeClassFromJacocoGeneratedReport;
-import kozlov.kirill.snake.view.SceneManager;
-
-@ExcludeClassFromJacocoGeneratedReport
-public interface SceneManagerAccessible {
-    void setSceneManager(SceneManager manager);
-}
