--- task_2_3_1/src/main/java/kozlov/kirill/snake/view_model/settings/Validator.java
@@ -1,10 +1,0 @@
-package kozlov.kirill.snake.view_model.settings;
-
-import javafx.scene.control.TextInputControl;
-import javafx.scene.input.KeyEvent;
-import kozlov.kirill.snake.ExcludeClassFromJacocoGeneratedReport;
-
-@ExcludeClassFromJacocoGeneratedReport
-public interface Validator {
-    boolean isValid(TextInputControl field, KeyEvent event);
-}
