+++ task_2_3_1/src/main/java/kozlov/kirill/snake/viewmodel/GameOverViewModel.java
@@ -1,4 +1,4 @@
-package kozlov.kirill.snake.view_model;
+package kozlov.kirill.snake.viewmodel;
 
 import javafx.fxml.FXML;
 import javafx.scene.control.Label;
@@ -8,6 +8,9 @@ import kozlov.kirill.snake.model.game.GameModel;
 import kozlov.kirill.snake.view.SceneEnum;
 import kozlov.kirill.snake.view.SceneManager;
 
+/**
+ * Game over scene view-model class.
+ */
 @ExcludeClassFromJacocoGeneratedReport
 public class GameOverViewModel implements SceneManagerAccessible {
     private SceneManager sceneManager;
@@ -23,6 +26,11 @@ public class GameOverViewModel implements SceneManagerAccessible {
         );
     }
 
+    /**
+     * Back button handler.
+     * <br>
+     * Returns to MENU scene
+     */
     @FXML
     protected void back() {
         sceneManager.changeScene(SceneEnum.MENU);
