+++ task_2_3_1/src/main/java/kozlov/kirill/snake/viewmodel/settings/TypeEventProcessor.java
@@ -1,4 +1,4 @@
-package kozlov.kirill.snake.view_model.settings;
+package kozlov.kirill.snake.viewmodel.settings;
 
 import javafx.event.EventHandler;
 import javafx.scene.control.Label;
@@ -6,6 +6,9 @@ import javafx.scene.control.TextInputControl;
 import javafx.scene.input.KeyEvent;
 import kozlov.kirill.snake.ExcludeClassFromJacocoGeneratedReport;
 
+/**
+ * Extended event handler class.
+ */
 @ExcludeClassFromJacocoGeneratedReport
 public class TypeEventProcessor implements EventHandler<KeyEvent> {
     private final TextInputControl fxmlField;
@@ -14,6 +17,15 @@ public class TypeEventProcessor implements EventHandler<KeyEvent> {
     private final String errorMessage;
     private String previousValue;
 
+    /**
+     * Constructor.
+     *
+     * @param fxmlField FXML input field for validation
+     * @param initialValue initial field value
+     * @param validator implementation of Validator interface
+     * @param errorLabel label where will be showed error message
+     * @param errorMessage message for showing in case of validation failed
+     */
     public TypeEventProcessor(
         TextInputControl fxmlField, String initialValue,
         Validator validator,
