+++ task_2_3_1/src/main/java/module-info.java
@@ -10,11 +10,11 @@ module kozlov.kirill.snake {
     opens kozlov.kirill.snake to javafx.fxml;
     exports kozlov.kirill.snake.view;
     opens kozlov.kirill.snake.view to javafx.fxml;
-    exports kozlov.kirill.snake.view_model;
-    opens kozlov.kirill.snake.view_model to javafx.fxml;
+    exports kozlov.kirill.snake.viewmodel;
+    opens kozlov.kirill.snake.viewmodel to javafx.fxml;
 
     exports kozlov.kirill.util to com.fasterxml.jackson.databind;
     exports kozlov.kirill.snake.model.settings to com.fasterxml.jackson.databind;
-    exports kozlov.kirill.snake.view_model.settings;
-    opens kozlov.kirill.snake.view_model.settings to javafx.fxml;
+    exports kozlov.kirill.snake.viewmodel.settings;
+    opens kozlov.kirill.snake.viewmodel.settings to javafx.fxml;
 }
