+++ task_2_3_1/src/main/resources/kozlov/kirill/snake/game-over-scene.fxml
@@ -10,7 +10,7 @@
 
 <?import javafx.scene.layout.HBox?>
 <VBox alignment="CENTER" spacing="20.0" xmlns:fx="http://javafx.com/fxml"
-      fx:controller="kozlov.kirill.snake.view_model.GameOverViewModel">
+      fx:controller="kozlov.kirill.snake.viewmodel.GameOverViewModel">
     <padding>
         <Insets bottom="20.0" left="20.0" right="20.0" top="20.0"/>
     </padding>
